+++ lib/engram/persistence.rb
@@ -16,9 +16,9 @@ module Engram
       @store.add(record) if record
     end
 
-    def update(id:, record:)
+    def update(scope:, id:, record:)
       record = prepare(record)
-      @store.update(id: id, record: record) if record
+      @store.update(scope: scope, id: id, record: record) if record
     end
 
     private